requirements.txt
pyserial==3.2.1
python-can==4.2.2
